requirements.txt
fastapi
uvicorn
pydantic
httpx
pytest
python-dotenv
google-genai
